requirements.txt
allure-pytest==2.8.32
pytest-xdist==2.4.0